dr/requirements.txt
# Core Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
xgboost>=1.5.0

# GUI Libraries
customtkinter>=5.1.2
Pillow>=9.0.0

# Data Visualization
matplotlib>=3.4.0

# Machine Learning
tensorflow>=2.8.0
keras>=2.8.0

# Image Processing
opencv-python>=4.5.0

# Utilities
joblib>=1.1.0
tqdm>=4.62.0

# Optional - for enhanced visualizations
seaborn>=0.11.0 